+++ src/main/java/com/AirBnb/AirBnb/service/RoomServiceImpl.java
@@ -3,13 +3,16 @@ package com.AirBnb.AirBnb.service;
 import com.AirBnb.AirBnb.dto.RoomDto;
 import com.AirBnb.AirBnb.entity.Hotel;
 import com.AirBnb.AirBnb.entity.Room;
+import com.AirBnb.AirBnb.entity.User;
 import com.AirBnb.AirBnb.exception.ResourceNotFoundException;
+import com.AirBnb.AirBnb.exception.UnAuthorisedException;
 import com.AirBnb.AirBnb.repository.HotelRepository;
 import com.AirBnb.AirBnb.repository.InventoryRepository;
 import com.AirBnb.AirBnb.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -39,6 +42,11 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel=hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID:{}"+hotelId));
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorisedException("This user does not own this hotel with id:"+hotelId);
+        }
         Room room=modelMapper.map(roomDto,Room.class);
         room.setHotel(hotel);
         room= roomRepository.save(room);
@@ -55,6 +63,11 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel=hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID:{}"+hotelId));
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorisedException("This user does not own this hotel with id:"+hotelId);
+        }
         return hotel.getRooms()
                 .stream()
                 .map((element) -> modelMapper.map(element, RoomDto.class))
@@ -77,8 +90,12 @@ public class RoomServiceImpl implements RoomService{
         Room room=roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID:{}"+roomId));
-
-       inventoryService.deleteAllInventories(room);
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner()))
+        {
+            throw new UnAuthorisedException("This user does not own this room with id:"+roomId);
+        }
+        inventoryService.deleteAllInventories(room);
        roomRepository.deleteById(roomId);
     }
 }
